# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ClaimResult": {
        "additionalProperties": false,
        "description": "The synchronous claim branch, without asset paths.\n\nThe three enum fields alone are unreadable: \"applicability=undetermined,\ndecision=conditional\" tells a user nothing about what was established,\nwhat is missing or which page supports it. The reasoning the domain\nalready computed travels with them.",
        "properties": {
          "applicability": {
            "enum": [
              "applicable",
              "not_applicable",
              "undetermined"
            ],
            "title": "Applicability",
            "type": "string"
          },
          "blocks": {
            "default": [],
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Blocks",
            "type": "array"
          },
          "conditions": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Conditions",
            "type": "array"
          },
          "contradictions": {
            "default": [],
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Contradictions",
            "type": "array"
          },
          "convention": {
            "anyOf": [
              {
                "enum": [
                  "CIDE",
                  "ASCIDE"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Convention"
          },
          "decision": {
            "enum": [
              "resolved",
              "conditional",
              "undetermined",
              "not_assessed"
            ],
            "title": "Decision",
            "type": "string"
          },
          "facts": {
            "default": [],
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Facts",
            "type": "array"
          },
          "kind": {
            "const": "claim",
            "title": "Kind",
            "type": "string"
          },
          "missing_information": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Missing Information",
            "type": "array"
          },
          "party_ids": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Party Ids",
            "type": "array"
          },
          "rules_evaluated": {
            "default": [],
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Rules Evaluated",
            "type": "array"
          },
          "trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Id"
          },
          "trace_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Url"
          }
        },
        "required": [
          "kind",
          "applicability",
          "convention",
          "decision"
        ],
        "title": "ClaimResult",
        "type": "object"
      },
      "ClarificationResultBody": {
        "additionalProperties": false,
        "description": "The auto-router branch when the classifier asks for more input.",
        "properties": {
          "kind": {
            "const": "clarification",
            "title": "Kind",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "missing_fields": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Missing Fields",
            "type": "array"
          }
        },
        "required": [
          "kind",
          "message"
        ],
        "title": "ClarificationResultBody",
        "type": "object"
      },
...
      "EnvelopeRequest": {
        "additionalProperties": false,
        "description": "One unified request across the three modes.\n\n``mode`` selects the explicit flow; ``profile`` optionally\noverrides the composition-root default. The route validates that\n``clarifications`` is only supplied for ``claim`` mode.",
        "properties": {
          "clarifications": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clarifications"
          },
          "language": {
            "default": "es",
            "enum": [
              "es",
              "en"
            ],
            "title": "Language",
            "type": "string"
          },
          "mode": {
            "enum": [
              "question",
              "claim",
              "auto"
            ],
            "title": "Mode",
            "type": "string"
          },
          "profile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile"
          },
          "session_id": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "stream": {
            "default": false,
            "title": "Stream",
            "type": "boolean"
          },
          "text": {
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text",
          "mode"
        ],
        "title": "EnvelopeRequest",
        "type": "object"
      },
      "EnvelopeResponse": {
        "additionalProperties": false,
        "description": "The unified response shape across the three modes.",
        "properties": {
          "evidence": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/EvidenceItem"
            },
            "title": "Evidence",
            "type": "array"
          },
          "metadata": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "default": {},
            "title": "Metadata",
            "type": "object"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "requested_mode": {
            "enum": [
              "question",
              "claim",
              "auto"
            ],
            "title": "Requested Mode",
            "type": "string"
          },
          "resolved_mode": {
            "enum": [
              "question",
              "claim",
              "clarification"
            ],
            "title": "Resolved Mode",
            "type": "string"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QuestionResult"
              },
              {
                "$ref": "#/components/schemas/ClaimResult"
              },
              {
                "$ref": "#/components/schemas/ClarificationResultBody"
              }
            ],
            "title": "Result"
          }
        },
        "required": [
          "request_id",
          "requested_mode",
          "resolved_mode",
          "result"
        ],
        "title": "EnvelopeResponse",
        "type": "object"
      },
      "EvidenceItem": {
        "additionalProperties": false,
        "description": "One piece of evidence exposed in the envelope (no asset paths).\n\nField names mirror the canonical evidence schema\n(``schemas/evidence.py:31-40``): ``document_hash`` and ``pdf_page``.\nNo ``version`` field because ``PageEvidence`` does not carry one;\nthe document hash already pins the version.",
        "properties": {
          "delivery": {
            "default": "text",
            "enum": [
              "text",
              "image",
              "rule"
            ],
            "title": "Delivery",
            "type": "string"
          },
          "document_hash": {
            "title": "Document Hash",
            "type": "string"
          },
          "evidence_id": {
            "title": "Evidence Id",
            "type": "string"
          },
          "pdf_page": {
            "title": "Pdf Page",
            "type": "integer"
          }
        },
        "required": [
          "evidence_id",
          "document_hash",
          "pdf_page"
        ],
        "title": "EvidenceItem",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "QueryResolveRequest": {
        "additionalProperties": false,
        "description": "One user query for the auto router.",
        "properties": {
          "language": {
            "default": "es",
            "enum": [
              "es",
              "en"
            ],
            "title": "Language",
            "type": "string"
          },
          "session_id": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "text": {
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "QueryResolveRequest",
        "type": "object"
      },
      "QueryResolveResponse": {
        "additionalProperties": false,
        "description": "The selected branch plus the rationale and trace identifier.",
        "properties": {
          "decision": {
            "enum": [
              "question",
              "claim",
              "clarification_required"
            ],
            "title": "Decision",
            "type": "string"
          },
          "rationale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rationale"
          },
          "trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Id"
          }
        },
        "required": [
          "decision",
          "rationale",
          "trace_id"
        ],
        "title": "QueryResolveResponse",
        "type": "object"
      },
...
      "QuestionResult": {
        "additionalProperties": false,
        "description": "The synchronous question branch.",
        "properties": {
          "blocks": {
            "default": [],
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Blocks",
            "type": "array"
          },
          "kind": {
            "const": "question",
            "title": "Kind",
            "type": "string"
          },
          "status": {
            "enum": [
              "answered",
              "partial",
              "insufficient_evidence",
              "out_of_scope"
            ],
            "title": "Status",
            "type": "string"
          },
          "trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Id"
          },
          "trace_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Url"
          }
        },
        "required": [
          "kind",
          "status"
        ],
        "title": "QuestionResult",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Allianz CICOS Claims Intelligence",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/queries": {
      "post": {
        "operationId": "envelope_api_v1_queries_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnvelopeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Envelope",
        "tags": [
          "queries"
        ]
      }
    },
    "/api/v1/queries/resolve": {
      "post": {
        "operationId": "resolve_query_route_api_v1_queries_resolve_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryResolveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResolveResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve Query Route",
        "tags": [
          "queries"
        ]
      }
    },
    "/api/v1/queries/stream": {
      "post": {
        "operationId": "envelope_stream_api_v1_queries_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnvelopeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Envelope Stream",
        "tags": [
          "queries"
        ]
      }
    },
...
  }
}